+++ BoltOn.sln
@@ -16,8 +16,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoltOn.Tests", "tests\BoltO
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BoltOn", "src\BoltOn\BoltOn.csproj", "{444D6887-C695-489B-9706-D3AB391482F4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BoltOn.Mediator.Data.EF", "src\BoltOn.Mediator.Data.EF\BoltOn.Mediator.Data.EF.csproj", "{E235DE19-202F-41E2-847F-1A34F8B3A984}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BoltOn.Data.EF", "src\BoltOn.Data.EF\BoltOn.Data.EF.csproj", "{E9FBB8EF-7628-4351-B008-A54EA28C67C2}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoltOn.Samples.DTOs", "samples\BoltOn.Samples.DTOs\BoltOn.Samples.DTOs.csproj", "{E659CB72-8D5A-43F4-91A5-D05D000C396B}"
@@ -44,10 +42,6 @@ Global
 		{444D6887-C695-489B-9706-D3AB391482F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{444D6887-C695-489B-9706-D3AB391482F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{444D6887-C695-489B-9706-D3AB391482F4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E235DE19-202F-41E2-847F-1A34F8B3A984}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E235DE19-202F-41E2-847F-1A34F8B3A984}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E235DE19-202F-41E2-847F-1A34F8B3A984}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E235DE19-202F-41E2-847F-1A34F8B3A984}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E9FBB8EF-7628-4351-B008-A54EA28C67C2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E9FBB8EF-7628-4351-B008-A54EA28C67C2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E9FBB8EF-7628-4351-B008-A54EA28C67C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -65,7 +59,6 @@ Global
 		{D5E25819-7EA3-4B05-B95D-BEE3EB9A18AB} = {15E5CCAC-85EB-4C98-B1FA-B5DEFD213FCD}
 		{6AB42094-D912-4DBE-8253-86409EA08896} = {C751203D-F070-44BC-88D6-C1687D30470F}
 		{444D6887-C695-489B-9706-D3AB391482F4} = {CF6C3199-B58A-4FB2-8E69-4E27B3A7BA29}
-		{E235DE19-202F-41E2-847F-1A34F8B3A984} = {CF6C3199-B58A-4FB2-8E69-4E27B3A7BA29}
 		{E9FBB8EF-7628-4351-B008-A54EA28C67C2} = {CF6C3199-B58A-4FB2-8E69-4E27B3A7BA29}
 		{E659CB72-8D5A-43F4-91A5-D05D000C396B} = {15E5CCAC-85EB-4C98-B1FA-B5DEFD213FCD}
 	EndGlobalSection
