+++ src/BoltOn.Data.EF/DbContextFactory.cs
@@ -11,15 +11,23 @@ namespace BoltOn.Data.EF
 	public class DbContextFactory : IDbContextFactory
 	{
 		private readonly IServiceProvider _serviceProvider;
+		readonly MediatorDataContext _mediatorDataContext;
 
-		public DbContextFactory(IServiceProvider serviceProvider)
+		public DbContextFactory(IServiceProvider serviceProvider,
+			MediatorDataContext mediatorDataContext)
 		{
+			_mediatorDataContext = mediatorDataContext;
 			_serviceProvider = serviceProvider;
 		}
 
 		public TDbContext Get<TDbContext>() where TDbContext : DbContext
 		{
 			var dbContext = _serviceProvider.GetService(typeof(TDbContext)) as TDbContext;
+			if (_mediatorDataContext.IsQueryRequest)
+			{
+				dbContext.ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
+				dbContext.ChangeTracker.AutoDetectChangesEnabled = false;
+			}
 			return dbContext;
 		}
 	}
