+++ src/BoltOn.Data.EF/RegistrationTask.cs
@@ -5,12 +5,11 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace BoltOn.Mediator.Data.EF
 {
-	public class MediatorDataRegistrationTask : IBootstrapperRegistrationTask
+	public class RegistrationTask : IRegistrationTask
 	{
 		public void Run(RegistrationTaskContext context)
 		{
 			var container = context.Container;
-			container.AddTransient<IDbContextFactory, MediatorDbContextFactory>();
 			container.AddScoped<MediatorDataContext>();
 			container.AddTransient<IInterceptor, EFQueryTrackingBehaviorInterceptor>();
 		}
