--- src/BoltOn.Mediator.Data.EF/MediatorDbContextFactory.cs
@@ -1,30 +1,0 @@
-using System;
-using BoltOn.Data.EF;
-using Microsoft.EntityFrameworkCore;
-
-namespace BoltOn.Mediator.Data.EF
-{
-	public class MediatorDbContextFactory : IDbContextFactory
-	{
-		private readonly IServiceProvider _serviceProvider;
-		private readonly MediatorDataContext _mediatorDataContext;
-
-		public MediatorDbContextFactory(IServiceProvider serviceProvider,
-			MediatorDataContext mediatorDataContext)
-		{
-			_serviceProvider = serviceProvider;
-			_mediatorDataContext = mediatorDataContext;
-		}
-
-		public TDbContext Get<TDbContext>() where TDbContext : DbContext
-		{
-			var dbContext = _serviceProvider.GetService(typeof(TDbContext)) as TDbContext;
-			if (_mediatorDataContext.IsQueryRequest)
-			{
-				dbContext.ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
-				dbContext.ChangeTracker.AutoDetectChangesEnabled = false;
-			}
-			return dbContext;
-		}
-	}
-}
