+++ tests/BoltOn.Tests/Mediator/MediatorTests.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Transactions;
 using BoltOn.Bootstrapping;
+using BoltOn.Data.EF;
 using BoltOn.Logging;
 using BoltOn.Mediator.Interceptors;
 using BoltOn.Mediator.Pipeline;
@@ -268,6 +269,96 @@ namespace BoltOn.Tests.Mediator
 			Assert.Equal(string.Format(Constants.ExceptionMessages.
 									   HANDLER_NOT_FOUND, request), result.Message);
    		}
+
+		[Fact]
+		public void Get_MediatorWithQueryRequest_ExecutesEFQueryTrackingBehaviorInterceptorAndDisablesTracking()
+		{
+			// arrange
+			var autoMocker = new AutoMocker();
+			var serviceProvider = autoMocker.GetMock<IServiceProvider>();
+			var testHandler = new Mock<TestHandler>();
+			var interceptor = new Mock<IInterceptor>();
+			var logger = new Mock<IBoltOnLogger<EFQueryTrackingBehaviorInterceptor>>();
+			serviceProvider.Setup(s => s.GetService(typeof(IRequestHandler<TestQuery, bool>)))
+				.Returns(testHandler.Object);
+			var request = new TestQuery();
+			var mediatorContext = new MediatorDataContext();
+			autoMocker.Use<IEnumerable<IInterceptor>>(new List<IInterceptor>
+			{
+				new EFQueryTrackingBehaviorInterceptor(logger.Object, mediatorContext)
+			});
+			var sut = autoMocker.CreateInstance<BoltOn.Mediator.Pipeline.Mediator>();
+			testHandler.Setup(s => s.Handle(request)).Returns(true);
+
+			// act
+			var result = sut.Process(request);
+
+			// assert 
+			Assert.True(result);
+			Assert.True(mediatorContext.IsQueryRequest);
+			logger.Verify(l => l.Debug($"Entering {nameof(EFQueryTrackingBehaviorInterceptor)}..."));
+			logger.Verify(l => l.Debug($"IsQueryRequest: {true}"));
+		}
+
+		[Fact]
+		public void Get_MediatorWithStaleQueryRequest_ExecutesEFQueryTrackingBehaviorInterceptorAndDisablesTracking()
+		{
+			// arrange
+			var autoMocker = new AutoMocker();
+			var serviceProvider = autoMocker.GetMock<IServiceProvider>();
+			var testHandler = new Mock<TestHandler>();
+			var interceptor = new Mock<IInterceptor>();
+			var logger = new Mock<IBoltOnLogger<EFQueryTrackingBehaviorInterceptor>>();
+			serviceProvider.Setup(s => s.GetService(typeof(IRequestHandler<TestStaleQuery, bool>)))
+				.Returns(testHandler.Object);
+			var request = new TestStaleQuery();
+			var mediatorContext = new MediatorDataContext();
+			autoMocker.Use<IEnumerable<IInterceptor>>(new List<IInterceptor>
+			{
+				new EFQueryTrackingBehaviorInterceptor(logger.Object, mediatorContext)
+			});
+			var sut = autoMocker.CreateInstance<BoltOn.Mediator.Pipeline.Mediator>();
+			testHandler.Setup(s => s.Handle(request)).Returns(true);
+
+			// act
+			var result = sut.Process(request);
+
+			// assert 
+			Assert.True(result);
+			Assert.True(mediatorContext.IsQueryRequest);
+			logger.Verify(l => l.Debug($"Entering {nameof(EFQueryTrackingBehaviorInterceptor)}..."));
+			logger.Verify(l => l.Debug($"IsQueryRequest: {true}"));
+		}
+
+		[Fact]
+		public void Get_MediatorWithCommandRequest_ExecutesEFQueryTrackingBehaviorInterceptorAndEnablesTracking()
+		{
+			// arrange
+			var autoMocker = new AutoMocker();
+			var serviceProvider = autoMocker.GetMock<IServiceProvider>();
+			var testHandler = new Mock<TestHandler>();
+			var interceptor = new Mock<IInterceptor>();
+			var logger = new Mock<IBoltOnLogger<EFQueryTrackingBehaviorInterceptor>>();
+			serviceProvider.Setup(s => s.GetService(typeof(IRequestHandler<TestCommand, bool>)))
+				.Returns(testHandler.Object);
+			var request = new TestCommand();
+			var mediatorContext = new MediatorDataContext();
+			autoMocker.Use<IEnumerable<IInterceptor>>(new List<IInterceptor>
+			{
+				new EFQueryTrackingBehaviorInterceptor(logger.Object, mediatorContext)
+			});
+			var sut = autoMocker.CreateInstance<BoltOn.Mediator.Pipeline.Mediator>();
+			testHandler.Setup(s => s.Handle(request)).Returns(true);
+
+			// act
+			var result = sut.Process(request);
+
+			// assert 
+			Assert.True(result);
+			Assert.False(mediatorContext.IsQueryRequest);
+			logger.Verify(l => l.Debug($"Entering {nameof(EFQueryTrackingBehaviorInterceptor)}..."));
+			logger.Verify(l => l.Debug($"IsQueryRequest: {false}"));
+		}
 
 		public void Dispose()
 		{
